refactor: extract the pty spawn functions, with deps injected (#548 step 3c-2)
The last and most entangled of the PTY machinery leaves index.ts: spawnClaudePty,
spawnCodexPty, spawnCommandPty, spawnLauncherPty and the codex relay/rollout
helpers. One module per kind of thing being started, because the claude path
(sandbox, CLI args, optimistic sidebar row, draft injection) has nothing in common
with running a shell command.

What index.ts still owns arrives as SpawnDeps rather than as imports: reap and
setWorking drive a session lifecycle far wider than spawning, and the hook/MCP
json is built from config it holds. Everything already extracted — the registry,
pty-spawn, draft-injection, ws-frames — is imported directly.

Every moved body is identical to its original apart from those injections.
